FilterGraph: minor clean into FrameBuffer class

+++ src/AvTranscoder/filter/FilterGraph.cpp
@@ -32,7 +32,7 @@ FrameBuffer::FrameBuffer(const AudioFrameDesc& audioFrameDesc)
 
 FrameBuffer::~FrameBuffer()
 {
-    for (int i = 0; i < _frameQueue.size(); ++i)
+    for(size_t i = 0; i < _frameQueue.size(); ++i)
         popFrame();
 }
